+++ src/components/VariablePanel/VariablePanel.test.tsx
@@ -1,21 +1,18 @@
 import React from 'react';
-import { locationService } from '@grafana/runtime';
-import { act, render, screen } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import { TestIds } from '../../constants';
+import { useTable } from '../../hooks';
 import { VariablePanel } from './VariablePanel';
 
 /**
- * Mock @grafana/runtime
+ * Mock hooks
  */
-jest.mock('@grafana/runtime', () => ({
-  ...jest.requireActual('@grafana/runtime'),
-  locationService: {
-    partial: jest.fn(),
-    getSearch: jest.fn(() => new URLSearchParams()),
-    getSearchObject: jest.fn(() => ({})),
-  },
-  getTemplateSrv: jest.fn(() => ({
-    getVariables: jest.fn(() => []),
+jest.mock('../../hooks', () => ({
+  ...jest.requireActual('../../hooks'),
+  useTable: jest.fn(() => ({
+    tableData: [],
+    columns: [],
+    getSubRows: () => undefined,
   })),
 }));
 
@@ -23,10 +20,6 @@ jest.mock('@grafana/runtime', () => ({
  * Panel
  */
 describe('Panel', () => {
-  beforeEach(() => {
-    jest.mocked(locationService).partial.mockClear();
-  });
-
   const eventBus = {
     getStream: jest.fn(() => ({
       subscribe: jest.fn(() => ({
@@ -42,23 +35,154 @@ describe('Panel', () => {
     return <VariablePanel width={100} height={100} eventBus={eventBus} {...restProps} options={options} />;
   };
 
-  /**
-   * Render without errors
-   * @param component
-   */
-  const renderWithoutErrors = async <T,>(component: React.ReactElement): Promise<void> => {
-    await render(component);
-
-    await new Promise((resolve) => setTimeout(resolve));
-  };
-
   it('Should find component', async () => {
-    await act(() => renderWithoutErrors(getComponent({})));
+    render(getComponent({}));
     expect(screen.getByTestId(TestIds.panel.root)).toBeInTheDocument();
   });
 
   it('Should show info message if no variables', async () => {
-    await act(() => renderWithoutErrors(getComponent({})));
+    render(getComponent({}));
     expect(screen.getByTestId(TestIds.panel.infoMessage)).toBeInTheDocument();
+  });
+
+  it('Should use first group', () => {
+    render(
+      getComponent({
+        options: {
+          groups: [
+            {
+              name: 'group1',
+              items: [
+                {
+                  name: 'group1Field',
+                },
+              ],
+            },
+            {
+              name: 'group2',
+              items: [],
+            },
+          ],
+        },
+      })
+    );
+
+    expect(useTable).toHaveBeenCalledWith(
+      expect.objectContaining({
+        levels: [
+          {
+            name: 'group1Field',
+          },
+        ],
+      })
+    );
+  });
+
+  it('Should use first group if already selected group removed', () => {
+    const { rerender } = render(
+      getComponent({
+        options: {
+          groups: [
+            {
+              name: 'group1',
+              items: [
+                {
+                  name: 'group1Field',
+                },
+              ],
+            },
+            {
+              name: 'group2',
+              items: [],
+            },
+          ],
+        },
+      })
+    );
+
+    expect(useTable).toHaveBeenCalledWith(
+      expect.objectContaining({
+        levels: [
+          {
+            name: 'group1Field',
+          },
+        ],
+      })
+    );
+
+    rerender(
+      getComponent({
+        options: {
+          groups: [
+            {
+              name: 'group2',
+              items: [
+                {
+                  name: 'group2Field',
+                },
+              ],
+            },
+          ],
+        },
+      })
+    );
+
+    expect(useTable).toHaveBeenCalledWith(
+      expect.objectContaining({
+        levels: [
+          {
+            name: 'group2Field',
+          },
+        ],
+      })
+    );
+  });
+
+  it('Should switch groups', () => {
+    jest.mocked(useTable).mockImplementation(() => ({
+      tableData: [{ value: 'device1', selected: false, showStatus: false }],
+      columns: [{ id: 'value', accessorKey: 'value' }],
+      getSubRows: () => undefined,
+    }));
+
+    render(
+      getComponent({
+        options: {
+          groups: [
+            {
+              name: 'group1',
+              items: [
+                {
+                  name: 'group1Field',
+                },
+              ],
+            },
+            {
+              name: 'group2',
+              items: [
+                {
+                  name: 'group2Field',
+                },
+              ],
+            },
+          ],
+        },
+      })
+    );
+
+    /**
+     * Select group2
+     */
+    fireEvent.click(screen.getByTestId(TestIds.panel.tab('group2')));
+
+    expect(useTable).toHaveBeenCalledWith(
+      expect.objectContaining({
+        levels: [
+          {
+            name: 'group2Field',
+          },
+        ],
+      })
+    );
   });
 });