+++ src/utils/table.ts
@@ -1,7 +1,7 @@
 import { DataFrame, Field, PanelData } from '@grafana/data';
 import { locationService } from '@grafana/runtime';
 import { FilterFn } from '@tanstack/react-table';
-import { Level, RuntimeVariable, TableItem } from '../../types';
+import { Level, RuntimeVariable, TableItem } from '../types';
 
 /**
  * Convert Data Frame to objects array
@@ -111,6 +111,7 @@ export const getRows = (
     value: item[key as keyof typeof item],
     selected: false,
     showStatus: false,
+    selectable: true,
   });
 
   return getGroupArray(