+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - Add an option to display the variable name in front of the value (#30)
 - Increase tests coverage (#31)
 - Add auto scroll to the selected value (#32)
-- Community signed release (#33)
+- Community signed release (#37)
 
 ## 1.3.0 (2023-06-30)